pyproject.toml
[project]
name = "radio-calisthenics-together"
version = "0.1.0"
description = "Radio Calisthenics YouTube Live Automator"
authors = [
    {name = "User", email = "user@example.com"}
]
dependencies = [
    "obsws-python>=1.7.0",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.9"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
